replaced "able-bodied" usersnap #4137
Student said: "The use of "able-bodied" to describe "non-disabled" people is considered a microaggression to community of people with disabilities.   This term is used in the body of this lesson."

I found here: https://ncdj.org/style-guide/
"Background: This term is used to describe someone who does not identify as having a disability. Some members of the disability community oppose its use because it implies that all people with disabilities lack “able bodies” or the ability to use their bodies well. They may prefer “non-disabled” or “enabled” as being more accurate.

NCDJ Recommendation: The term “non-disabled” and the phrases “does not have a disability” or “is not living with a disability” are more neutral choices. “Able-bodied” is an appropriate term to use in some cases, such as when referring to government reports on the proportion of able-bodied members in the workforce. In some cases, the word “typical” can be used to describe a non-disabled condition, although be aware that some in the disability community object to its use."